moving interpreter functions to their appropriate plugins (#33552)
