+++ src/dev/jest/config.js
@@ -44,6 +44,7 @@ export default {
     '!packages/kbn-ui-framework/src/services/**/*/index.js',
   ],
   moduleNameMapper: {
+    '^plugins/(.*)/(.*)': '<rootDir>/src/legacy/core_plugins/$1/public/$2',
     '^ui/(.*)': '<rootDir>/src/legacy/ui/public/$1',
     '^uiExports/(.*)': '<rootDir>/src/dev/jest/mocks/file_mock.js',
     '^test_utils/(.*)': '<rootDir>/src/test_utils/public/$1',
