+++ src/legacy/core_plugins/input_control_vis/index.js
@@ -25,6 +25,7 @@ export default function (kibana) {
       visTypes: [
         'plugins/input_control_vis/register_vis'
       ],
+      interpreter: ['plugins/input_control_vis/input_control_fn'],
       styleSheetPaths: resolve(__dirname, 'public/index.scss'),
     }
   });
