+++ src/legacy/core_plugins/metric_vis/index.js
@@ -27,6 +27,7 @@ export default function (kibana) {
       visTypes: [
         'plugins/metric_vis/metric_vis'
       ],
+      interpreter: ['plugins/metric_vis/metric_vis_fn'],
       styleSheetPaths: resolve(__dirname, 'public/index.scss'),
     }
