+++ src/legacy/core_plugins/metrics/index.js
@@ -31,6 +31,7 @@ export default function (kibana) {
       visTypes: [
         'plugins/metrics/kbn_vis_types',
       ],
+      interpreter: ['plugins/metrics/tsvb_fn'],
       styleSheetPaths: resolve(__dirname, 'public/index.scss'),
     },
