+++ src/legacy/core_plugins/region_map/index.js
@@ -24,6 +24,7 @@ export default function (kibana) {
   return new kibana.Plugin({
     uiExports: {
       visTypes: ['plugins/region_map/region_map_vis'],
+      interpreter: ['plugins/region_map/region_map_fn'],
       styleSheetPaths: resolve(__dirname, 'public/index.scss'),
     }
   });
