+++ src/legacy/core_plugins/region_map/public/region_map_fn.js
@@ -17,34 +17,42 @@
  * under the License.
  */
 
-import { functionWrapper } from '../../test_helpers';
-import { tagcloud } from './tagcloud';
+import { functionsRegistry } from 'plugins/interpreter/registries';
+import { i18n } from '@kbn/i18n';
 
-describe('interpreter/functions#tagcloud', () => {
-  const fn = functionWrapper(tagcloud);
-  const context = {
-    type: 'kibana_datatable',
-    rows: [{ 'col-0-1': 0 }],
-    columns: [{ id: 'col-0-1', name: 'Count' }],
-  };
-  const visConfig = {
-    scale: 'linear',
-    orientation: 'single',
-    minFontSize: 18,
-    maxFontSize: 72,
-    showLabel: true,
-    metric: {
-      accessor: 0,
-      format: {
-        id: 'number',
-      },
-      params: {},
-      aggType: 'count',
+export const regionmap = () => ({
+  name: 'regionmap',
+  type: 'render',
+  context: {
+    types: [
+      'kibana_datatable'
+    ],
+  },
+  help: i18n.translate('regionMap.function.help', {
+    defaultMessage: 'Regionmap visualization'
+  }),
+  args: {
+    visConfig: {
+      types: ['string', 'null'],
+      default: '"{}"',
     },
-  };
+  },
+  fn(context, args) {
+    const visConfig = JSON.parse(args.visConfig);
 
-  it('returns an object with the correct structure', () => {
-    const actual = fn(context, { visConfig: JSON.stringify(visConfig) });
-    expect(actual).toMatchSnapshot();
-  });
+    return {
+      type: 'render',
+      as: 'visualization',
+      value: {
+        visData: context,
+        visType: 'region_map',
+        visConfig,
+        params: {
+          listenOnChange: true,
+        }
+      },
+    };
+  },
 });
+
+functionsRegistry.register(regionmap);
