+++ src/legacy/core_plugins/table_vis/index.js
@@ -26,6 +26,7 @@ export default function (kibana) {
       visTypes: [
         'plugins/table_vis/table_vis'
       ],
+      interpreter: ['plugins/table_vis/table_vis_fn'],
       styleSheetPaths: resolve(__dirname, 'public/index.scss'),
     },
   });
