+++ src/legacy/core_plugins/table_vis/public/table_vis_fn.js
@@ -17,19 +17,20 @@
  * under the License.
  */
 
-import { convertToGeoJson } from 'ui/vis/map/convert_to_geojson';
+import { functionsRegistry } from 'plugins/interpreter/registries';
+import { LegacyResponseHandlerProvider as legacyResponseHandlerProvider } from 'ui/vis/response_handlers/legacy';
 import { i18n } from '@kbn/i18n';
 
-export const tilemap = () => ({
-  name: 'tilemap',
+export const kibanaTable = () => ({
+  name: 'kibana_table',
   type: 'render',
   context: {
     types: [
       'kibana_datatable'
     ],
   },
-  help: i18n.translate('interpreter.functions.tilemap.help', {
-    defaultMessage: 'Tilemap visualization'
+  help: i18n.translate('tableVis.function.help', {
+    defaultMessage: 'Table visualization'
   }),
   args: {
     visConfig: {
@@ -37,22 +38,18 @@ export const tilemap = () => ({
       default: '"{}"',
     },
   },
-  fn(context, args) {
+  async fn(context, args) {
     const visConfig = JSON.parse(args.visConfig);
 
-    const { geohash, metric, geocentroid } = visConfig.dimensions;
-    const convertedData = convertToGeoJson(context, {
-      geohash,
-      metric,
-      geocentroid,
-    });
+    const responseHandler = legacyResponseHandlerProvider().handler;
+    const convertedData = await responseHandler(context, visConfig.dimensions);
 
     return {
       type: 'render',
       as: 'visualization',
       value: {
         visData: convertedData,
-        visType: 'tile_map',
+        visType: 'table',
         visConfig,
         params: {
           listenOnChange: true,
@@ -61,3 +58,5 @@ export const tilemap = () => ({
     };
   },
 });
+
+functionsRegistry.register(kibanaTable);
