+++ src/legacy/core_plugins/table_vis/public/table_vis_fn.test.js
@@ -17,58 +17,46 @@
  * under the License.
  */
 
-import { functionWrapper } from '../../test_helpers';
-import { kibanaPie } from './pie';
+import { functionWrapper } from '../../interpreter/test_helpers';
+import { kibanaTable } from './table_vis_fn';
 
 const mockResponseHandler = jest.fn().mockReturnValue(Promise.resolve({
-  hits: 1,
-  names: ['Count'],
-  raw: {
-    columns: [],
-    rows: [],
-  },
-  slices: {
-    children: [],
-  },
-  tooltipFormatter: {
-    id: 'number',
-  },
+  tables: [{ columns: [], rows: [] }],
 }));
-jest.mock('ui/vis/response_handlers/vislib', () => ({
-  VislibSlicesResponseHandlerProvider: () => ({ handler: mockResponseHandler }),
+jest.mock('ui/vis/response_handlers/legacy', () => ({
+  LegacyResponseHandlerProvider: () => ({ handler: mockResponseHandler }),
 }));
 
-describe('interpreter/functions#pie', () => {
-  const fn = functionWrapper(kibanaPie);
+describe('interpreter/functions#table', () => {
+  const fn = functionWrapper(kibanaTable);
   const context = {
     type: 'kibana_datatable',
     rows: [{ 'col-0-1': 0 }],
     columns: [{ id: 'col-0-1', name: 'Count' }],
   };
   const visConfig = {
-    addTooltip: true,
-    addLegend: false,
-    type: 'pie',
-    addTooltip: true,
-    addLegend: true,
-    legendPosition: 'right',
-    isDonut: true,
-    labels: {
-      show: false,
-      values: true,
-      last_level: true,
-      truncate: 100,
+    perPage: 10,
+    showPartialRows: false,
+    showMetricsAtAllLevels: false,
+    sort: {
+      columnIndex: null,
+      direction: null
     },
+    showTotal: false,
+    totalFunc: 'sum',
     dimensions: {
-      metric: {
-        accessor: 0,
-        format: {
-          id: 'number',
-        },
-        params: {},
-        aggType: 'count',
-      },
-    },
+      metrics: [
+        {
+          accessor: 0,
+          format: {
+            id: 'number'
+          },
+          params: {},
+          aggType: 'count'
+        }
+      ],
+      buckets: []
+    }
   };
 
   beforeEach(() => {
